simple refactoring done, removed some dead code and split components
